+++ account-keeper-impl/src/main/java/com/dwarfeng/acckeeper/impl/bean/key/HibernateProtectorVariableKey.java
@@ -2,9 +2,11 @@ package com.dwarfeng.acckeeper.impl.bean.key;
 
 import com.dwarfeng.subgrade.stack.bean.key.Key;
 
+import java.util.Objects;
+
 public class HibernateProtectorVariableKey implements Key {
 
-    private static final long serialVersionUID = 2620113708077340102L;
+    private static final long serialVersionUID = 5192976103699990164L;
 
     private String protectorInfoId;
     private String variableId;
@@ -31,6 +33,25 @@ public class HibernateProtectorVariableKey implements Key {
 
     public void setVariableId(String variableId) {
         this.variableId = variableId;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+
+        HibernateProtectorVariableKey that = (HibernateProtectorVariableKey) o;
+
+        if (!Objects.equals(protectorInfoId, that.protectorInfoId))
+            return false;
+        return Objects.equals(variableId, that.variableId);
+    }
+
+    @Override
+    public int hashCode() {
+        int result = protectorInfoId != null ? protectorInfoId.hashCode() : 0;
+        result = 31 * result + (variableId != null ? variableId.hashCode() : 0);
+        return result;
     }
 
     @Override